@@ -1,24 +1,13 @@
 import 'dart:core';
-import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:flutter/foundation.dart'
-    show debugDefaultTargetPlatformOverride;
 
-import 'src/basic_sample/basic_sample.dart';
 import 'src/call_sample/call_sample.dart';
 import 'src/call_sample/data_channel_sample.dart';
 import 'src/route_item.dart';
 
-bool isDesktop() {
-  return Platform.isWindows || Platform.isLinux || Platform.isMacOS;
-}
-
-void main(){
- if(isDesktop())
-   debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
-  runApp(new MyApp());
-}
+void main() => runApp(new MyApp());
 
 class MyApp extends StatefulWidget {
   @override
@@ -31,9 +20,10 @@ enum DialogDemoAction {
 }
 
 class _MyAppState extends State<MyApp> {
-  List<RouteItem> items;
-  String _serverAddress = '';
-  SharedPreferences prefs;
+  List<RouteItem> items = [];
+  String _server = '';
+  late SharedPreferences _prefs;
+
   bool _datachannel = false;
   @override
   initState() {
@@ -55,12 +45,12 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
-          appBar: new AppBar(
-            title: new Text('Flutter-WebRTC example'),
+    return MaterialApp(
+      home: Scaffold(
+          appBar: AppBar(
+            title: Text('Flutter-WebRTC example'),
           ),
-          body: new ListView.builder(
+          body: ListView.builder(
               shrinkWrap: true,
               padding: const EdgeInsets.all(0.0),
               itemCount: items.length,
@@ -71,26 +61,28 @@ class _MyAppState extends State<MyApp> {
   }
 
   _initData() async {
-    prefs = await SharedPreferences.getInstance();
+    _prefs = await SharedPreferences.getInstance();
     setState(() {
-      _serverAddress = prefs.getString('server') ?? 'demo.cloudwebrtc.com';
+      _server = _prefs.getString('server') ?? 'demo.cloudwebrtc.com';
     });
   }
 
-  void showDemoDialog<T>({BuildContext context, Widget child}) {
+  void showDemoDialog<T>(
+      {required BuildContext context, required Widget child}) {
     showDialog<T>(
       context: context,
       builder: (BuildContext context) => child,
-    ).then<void>((T value) {
+    ).then<void>((T? value) {
       // The value passed to Navigator.pop() or null.
       if (value != null) {
         if (value == DialogDemoAction.connect) {
-          prefs.setString('server', _serverAddress);
+          _prefs.setString('server', _server);
           Navigator.push(
               context,
               MaterialPageRoute(
-                  builder: (BuildContext context) =>
-                      _datachannel? DataChannelSample(ip: _serverAddress) : CallSample(ip: _serverAddress)));
+                  builder: (BuildContext context) => _datachannel
+                      ? DataChannelSample(host: _server)
+                      : CallSample(host: _server)));
         }
       }
     });
@@ -99,26 +91,26 @@ class _MyAppState extends State<MyApp> {
   _showAddressDialog(context) {
     showDemoDialog<DialogDemoAction>(
         context: context,
-        child: new AlertDialog(
+        child: AlertDialog(
             title: const Text('Enter server address:'),
             content: TextField(
               onChanged: (String text) {
                 setState(() {
-                  _serverAddress = text;
+                  _server = text;
                 });
               },
               decoration: InputDecoration(
-                hintText: _serverAddress,
+                hintText: _server,
               ),
               textAlign: TextAlign.center,
             ),
             actions: <Widget>[
-              new FlatButton(
+              FlatButton(
                   child: const Text('CANCEL'),
                   onPressed: () {
                     Navigator.pop(context, DialogDemoAction.cancel);
                   }),
-              new FlatButton(
+              FlatButton(
                   child: const Text('CONNECT'),
                   onPressed: () {
                     Navigator.pop(context, DialogDemoAction.connect);
@@ -128,15 +120,6 @@ class _MyAppState extends State<MyApp> {
 
   _initItems() {
     items = <RouteItem>[
-      RouteItem(
-          title: 'Basic API Tests',
-          subtitle: 'Basic API Tests.',
-          push: (BuildContext context) {
-            Navigator.push(
-                context,
-                new MaterialPageRoute(
-                    builder: (BuildContext context) => new BasicSample()));
-          }),
       RouteItem(
           title: 'P2P Call Sample',
           subtitle: 'P2P Call Sample.',
