@@ -5,9 +5,9 @@ typedef void RouteCallback(BuildContext context);
 
 class RouteItem {
   RouteItem({
-    @required this.title,
-    @required this.subtitle,
-    @required this.push,
+    required this.title,
+    required this.subtitle,
+    required this.push,
   });
 
   final String title;
