@@ -5,7 +5,7 @@ class CallSettings extends StatefulWidget {
   static String tag = 'call_settings';
 
   @override
-  _CallSettingsState createState() => new _CallSettingsState();
+  _CallSettingsState createState() => _CallSettingsState();
 }
 
 class _CallSettingsState extends State<CallSettings> {
@@ -17,20 +17,17 @@ class _CallSettingsState extends State<CallSettings> {
   @override
   deactivate() {
     super.deactivate();
-
   }
 
   @override
   Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        title: new Text('Settings'),
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Settings'),
       ),
-      body: new OrientationBuilder(
+      body: OrientationBuilder(
         builder: (context, orientation) {
-          return new Center(
-            child: Text("settings")
-          );
+          return Center(child: Text("settings"));
         },
       ),
     );
